+++ src/Ext/Rules/Body.cpp
@@ -451,10 +451,14 @@ void RulesExt::ExtData::LoadBeforeTypeData(RulesClass* pThis, CCINIClass* pINI)
 	this->AutoRemoveEarliestBeacon.Read(exINI, GameStrings::General, "AutoRemoveEarliestBeacon");
 	this->AllowBeaconHotKeyInSinglePlayer.Read(exINI, GameStrings::General, "AllowBeaconHotKeyInSinglePlayer");
 
-	this->ShowPassengers.Read(exINI, GameStrings::AudioVisual, "ShowPassengers");
-	this->ShowPassengers_Toggleable.Read(exINI, GameStrings::AudioVisual, "ShowPassengers.Toggleable");
-	this->ShowPassengers_PerRow.Read(exINI, GameStrings::AudioVisual, "ShowPassengers.PerRow");
-	this->ShowPassengers_BottomOffset.Read(exINI, GameStrings::AudioVisual, "ShowPassengers.BottomOffset");
+	this->ShowCameo.Read(exINI, GameStrings::AudioVisual, "ShowCameo");
+	this->ShowCameo_Toggleable.Read(exINI, GameStrings::AudioVisual, "ShowCameo.Toggleable");
+	this->ShowCameo_PerRow.Read(exINI, GameStrings::AudioVisual, "ShowCameo.PerRow");
+	this->ShowCameo_BottomOffset.Read(exINI, GameStrings::AudioVisual, "ShowCameo.BottomOffset");
+	this->ShowCameo_Translucency.Read(exINI, GameStrings::AudioVisual, "ShowCameo.Translucency");
+	this->ShowCameo_OverlapXY.Read(exINI, GameStrings::AudioVisual, "ShowCameo.OverlapXY");
+	this->ShowCameo_OffsetXY.Read(exINI, GameStrings::AudioVisual, "ShowCameo.OffsetXY");
+	this->ShowCameo_OverlapPrimary.Read(exINI, GameStrings::AudioVisual, "ShowCameo.OverlapPrimary");
 
 	// Section AITargetTypes
 	int itemsCount = pINI->GetKeyCount("AITargetTypes");
@@ -805,10 +809,14 @@ void RulesExt::ExtData::Serialize(T& Stm)
 		.Process(this->SecondaryFireSequenceLandOnly)
 		.Process(this->AutoRemoveEarliestBeacon)
 		.Process(this->AllowBeaconHotKeyInSinglePlayer)
-		.Process(this->ShowPassengers)
-		.Process(this->ShowPassengers_Toggleable)
-		.Process(this->ShowPassengers_PerRow)
-		.Process(this->ShowPassengers_BottomOffset)
+		.Process(this->ShowCameo)
+		.Process(this->ShowCameo_Toggleable)
+		.Process(this->ShowCameo_PerRow)
+		.Process(this->ShowCameo_BottomOffset)
+		.Process(this->ShowCameo_Translucency)
+		.Process(this->ShowCameo_OverlapXY)
+		.Process(this->ShowCameo_OffsetXY)
+		.Process(this->ShowCameo_OverlapPrimary)
 		;
 }